+++ backend/env/unit.sh
@@ -44,6 +44,9 @@ export MC_FAKTORY_PORT=7419
 export MC_FAKTORY_NAME="mc-faktory-unit"
 
 export MC_API_URL="http://mcunit.localhost/api"
+# does not word as local nginx in not ssh
+# export MC_GLOBUS_AUTH_CALLBACK="$MC_API_URL/etl/globus/auth/callback"
+export MC_GLOBUS_AUTH_CALLBACK="https://localhost:5032/globus/auth/callback"
 
 export MC_ETL_WORKER_LOG_LEVEL=INFO