+++ backend/servers/etlserver/app/mc_etlserver.py
@@ -14,7 +14,8 @@ if not sys.path[0] == '':
 
 _HOST = os.environ.get('MC_SERVICE_HOST') or 'localhost'
 _PORT = os.environ.get('MC_ETL_SERVICE_PORT')
-
+_SSL_DIR = os.environ.get('MC_ETL_SSL_DIR')
+SERVER_TYPE = os.environ.get('SERVERTYPE')
 
 # noinspection PyUnusedLocal
 def reload_users(signum, frame):
@@ -34,7 +35,19 @@ def main():
         exit(-1)
 
     log.info("Starting ELT SERVER with host = {} and port = {}".format(_HOST, _PORT))
-    app.run(debug=True, host=_HOST, port=int(_PORT), processes=1)
+
+    if SERVER_TYPE and SERVER_TYPE == 'dev':
+        if not _SSL_DIR:
+            log.error("Environment missing MC_ETL_SSL_DIR; can not run server; quitting")
+
+        log.info("Starting https server using SSL dir - {}".format(_SSL_DIR))
+        app.run(debug=True, host=_HOST, port=int(_PORT),
+                ssl_context=(os.path.join(_SSL_DIR,'server.crt'),
+                             os.path.join(_SSL_DIR,'server.key')),
+                processes=1)
+    else:
+        log.info("Starting http server")
+        app.run(debug=True, host=_HOST, port=int(_PORT), processes=5, threaded=False)
 
 
 if __name__ == '__main__':