+++ backend/servers/etlserver/common/McdirHelper.py
@@ -11,17 +11,22 @@ class McdirHelper:
     def get_download_dir(self):
         return self._get_base_dir('__download_staging')
 
+    def base_dir(self):
+        return self._get_base_dir(None)
+
     @staticmethod
     def _get_base_dir(child):
         mcdir = os.environ.get('MCDIR')
         if not mcdir:
-            raise GlobusSetupException("$MDDIR is None - must be defined")
+            raise GlobusSetupException("$MCDIR is None - must be defined")
         base_path = mcdir.split(':')[0]
         if not os.path.exists(base_path):
             raise GlobusSetupException(
-                "the $MDDIR base path = {} - does not exist".format(base_path))
-        child_dir = os.path.join(base_path, child)
-        if not os.path.exists(child_dir):
-            os.mkdir(child_dir)
-        return child_dir
-
+                "the $MCDIR base path = {} - does not exist".format(base_path))
+        if child:
+            child_dir = os.path.join(base_path, child)
+            if not os.path.exists(child_dir):
+                os.mkdir(child_dir)
+            return child_dir
+        else:
+            return base_path