+++ website/src/app/layout/index.module.js
@@ -1,2 +1,3 @@
 import './navbar/navbar.directive';
 import './sidenav/mc-project-sidenav.component';
+import './sidenav/sidenav-globus.service';